Move mapCMSOptionsToListboxFormat to shared UI library
