+++ apps/ff-site/src/app/_utils/getCMSFieldOptionsAndValidIds.ts
@@ -2,11 +2,11 @@ import {
   getCMSFieldOptions,
   getCollectionConfig,
 } from '@filecoin-foundation/utils/cmsConfigUtils'
+import { mapCMSOptionsToListboxFormat } from '@filecoin-foundation/utils/mapCMSOptionsToListboxFormat'
 import type { CMSCollectionName } from '@filecoin-foundation/utils/types/cmsConfig'
 
 import configData from '@/data/cmsConfigSchema.json'
 
-import { mapCMSOptionsToListboxFormat } from '@/utils/mapCMSOptionsToListboxFormat'
 
 export type CMSFieldOptionsAndValidIdsParams = {
   collectionName: CMSCollectionName
